Module Lmdb__Lmdb_bindings.PutFlags

include Flags

type t

The type of a set of flags

val (+) : t -> t -> t

a + b is the union of flag sets a and b. This corresponds to a bitwise or on C bitfields.

val (*) : t -> t -> t

a * b is the intersection of flag sets a and b. This corresponds to a bitwise and on C bitfields.

val test : t -> t -> bool

test a b is true only if a is a subset of b. This corresponds to a & b == a for C bitfields.

val unset : t -> t -> t

unset a b removes flags a from flag set b. This corresponds to a & ~b for C bitfields.

val eq : t -> t -> bool

eq a b The equals relation.

val of_int : int -> t

val to_int : t -> int

val none : t

none The empty set of flags.

val no_overwrite : t

Raise Exists if the key already exists no matter whether the map supports duplicates.

val no_dup_data : t

Only for maps supporting duplicates: Don't add the value to an already existing key if this value is already part of this key.

val append : t

Add a key that is greater than any existing key. Used to efficiently add sorted data.

val append_dup : t

Add value to key that is greater than any existing value of this key. Used to efficiently add sorted values to a key.

val current : t

val reserve : t

val multiple : t